how the program  works
function prototypes

int reverse (int n); finds the reverse of a number.

int isPalindrome (int n);  finds if the reverse of a number is a palindrome then returns it.

int isPriime ( int number); Checks for all numbers which are prime.

int findSqr (int number); finds the square root of a number and returns it.

int IsSquare_of_prime (int  n); checks if the square root of a number is prime